Define failures using the iteration index.
This allows to get which specific item in an array does not match the rules.
A possible usecase is having to display an error message on the correct
nested form field.

At the moment, the failure key is something like "books:*:title". This commit
will change that to, for example, "books:4:title".